list files in new activity
