+++ app/src/main/res/values/strings.xml
@@ -31,6 +31,10 @@
     <string name="main_usb_open_failed">Could not open the selected storage.</string>
     <string name="main_usb_no_files">No files found on the selected drive.</string>
     <string name="main_usb_listing_prefix">Files on USB drive:</string>
+    <string name="usb_list_title">USB files</string>
+    <string name="usb_list_loading">Loading files...</string>
+    <string name="usb_list_no_uri">No USB location was provided.</string>
+    <string name="usb_list_error">Could not open the selected storage.</string>
 
 
     <!-- Strings used on EditActivity -->
